+++ CHANGES.rst
@@ -5,4 +5,6 @@ Changelog of lizard-geo
 0.1 (unreleased)
 ----------------
 
-- Initial library skeleton created by nensskel.  [your name]
+- Moved models from lizard-area.
+
+- Initial library skeleton created by nensskel.  [Jack Ha]